(wip) refactor FSRS struct
